requirements.txt
Flask>=3.0
Flask-Login>=0.6
Flask-WTF>=1.2
WTForms>=3.1
SQLAlchemy>=2.0
psycopg[binary]>=3.2
python-dotenv>=1.0
gunicorn>=21.2
bcrypt
